requirements.txt
openai>=1.37.0
anthropic>=0.30.0
httpx>=0.27.0
PyYAML>=6.0.1
pandas>=2.1.0
tqdm>=4.66.0
tenacity>=8.2.3
regex>=2024.5.15
python-dotenv>=1.0.1
